API/requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==4.4.2
deprecation==2.1.0
executing==1.2.0
ffmpeg==1.4
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gotrue==1.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
haversine==2.8.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
multidict==6.0.4
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
postgrest==0.11.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
pyzmq==25.1.1
realtime==1.0.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
Rtree==1.1.0
scikit-image==0.21.0
scikit-learn==1.3.1
scikit-surprise==1.1.3
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.30
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
websockets==10.4
Werkzeug==3.0.0
wrapt==1.14.1
yarl==1.9.2
youtube-dl==2021.12.17
youtube-search-python==1.6.6
